squash: follow review hints
